requirements.txt
pydantic~=2.10.6
fastapi~=0.115.8
SQLAlchemy~=2.0.37
pydantic-settings~=2.7.1
alembic~=1.14.1
PyJWT~=2.10.1
passlib~=1.7.4
asyncpg~=0.30.0
psycopg2~=2.9.10